requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.4
markupsafe==2.1.5
packaging==24.1
pip==24.1.2
pluggy==1.5.0
pytest==8.2.2
python-dotenv==1.0.1
ruff==0.5.1
sqlalchemy==2.0.31
typing-extensions==4.12.2
uv==0.2.24
werkzeug==3.0.3
